Scope method creation
